pyproject.toml
[tool.poetry]
name = "utils"
version = "0.1.0"
description = "Projeto com funções úteis para o dia a dia"
authors = ["Marco Antônio Martins Porto Netto <marco.netto@mjv.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
streamlit = "^1.30.0"
streamlit-option-menu = "^0.3.12"
pillow = "^10.2.0"
rembg = "^2.0.53"
pytube = "^15.0.0"


[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
isort = "^5.13.2"
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
ruff = "^0.1.14"
taskipy = "^1.12.2"

[tool.ruff]
line-length = 79
exclude = ['.venv', 'migrations']

[tool.isort]
profile = "black"
line_length = 79
extend_skip = ['migrations']

[tool.pytest.ini_options]
pythonpath = "."

[tool.blue]
extend-exclude = '(migrations/)'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
lint = 'ruff . && blue --check . --diff'
format = 'blue .  && isort .'
run = "streamlit run app/main.py"
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_zero -vv'
post_test = 'coverage html'
